Pull CanFrame out of candump module
