+++ src/can/candump.rs
@@ -2,7 +2,8 @@
 use std::io::{BufRead, BufReader, Lines, Read};
 
 use eyre::WrapErr;
-use serde::ser::SerializeStruct;
+
+use crate::can::CanFrame;
 
 /// File format of the candump
 #[derive(Clone, Copy, Debug, PartialEq)]
@@ -16,91 +17,6 @@ pub enum CandumpFormat {
     CanUtilsFile,
     /// candump -ta format
     CanUtilsCli,
-}
-
-#[derive(Clone, Debug, Default, PartialEq)]
-pub struct CanFrame {
-    pub timestamp: f64,
-    pub interface: String,
-    pub canid: u32,
-    pub dlc: usize,
-
-    data: [u8; 8],
-}
-
-impl CanFrame {
-    #[must_use]
-    pub fn data(&self) -> &[u8] {
-        &self.data[..self.dlc]
-    }
-
-    #[must_use]
-    pub fn dst(&self) -> u8 {
-        if self.is_point_to_point() {
-            self.pdu_specific() as u8
-        } else {
-            0xFF // global
-        }
-    }
-
-    #[must_use]
-    pub fn src(&self) -> u8 {
-        (self.canid & 0xFF) as u8
-    }
-
-    #[must_use]
-    pub fn priority(&self) -> u8 {
-        let shifted = self.canid >> 26;
-        let masked = shifted & 0b0111;
-        masked as u8
-    }
-
-    #[must_use]
-    pub fn is_point_to_point(&self) -> bool {
-        // destination-specific range is 00..=EF
-        // broadcast range is F0..=FF
-        self.pdu_format() <= 0xEF
-    }
-
-    #[must_use]
-    pub fn pdu_format(&self) -> u32 {
-        (self.canid & 0xFF0000) >> 16
-    }
-
-    #[must_use]
-    pub fn pdu_specific(&self) -> u32 {
-        (self.canid & 0x00FF00) >> 8
-    }
-
-    #[must_use]
-    pub fn pgn(&self) -> u32 {
-        // Shift off the src address
-        let canid = self.canid >> 8;
-        // Mask off the priority bits, leaving the EDP and DP data page bits
-        let canid = canid & 0x3FFFF;
-
-        if self.is_point_to_point() {
-            canid & 0x3FF00
-        } else {
-            canid
-        }
-    }
-}
-
-impl serde::Serialize for CanFrame {
-    fn serialize<S: serde::Serializer>(&self, serializer: S) -> Result<S::Ok, S::Error> {
-        let mut state = serializer.serialize_struct("CanFrame", 5)?;
-        state.serialize_field("timestamp", &self.timestamp)?;
-        state.serialize_field("interface", &self.interface)?;
-        state.serialize_field("canid", &format!("{:#X}", self.canid))?;
-        state.serialize_field("dlc", &self.dlc)?;
-        state.serialize_field("priority", &self.priority())?;
-        state.serialize_field("src", &format!("{:#X}", self.src()))?;
-        state.serialize_field("dst", &format!("{:#X}", self.dst()))?;
-        state.serialize_field("pgn", &format!("{:#X}", self.pgn()))?;
-        state.serialize_field("data", &hex::encode_upper(self.data()))?;
-        state.end()
-    }
 }
 
 /// Parse [CanFrame]s from the given reader
@@ -225,13 +141,13 @@ fn parse_candump_cli_msg(line: &str) -> eyre::Result<CanFrame> {
         let byte = u8::from_str_radix(maybe_byte, 16).wrap_err("Failed to parse data byte")?;
         data[i] = byte;
     }
-    Ok(CanFrame {
+    Ok(CanFrame::new(
         timestamp,
-        interface: interface.to_string(),
+        interface.to_string(),
         canid,
         dlc,
         data,
-    })
+    ))
 }
 
 /// Parse candumps with the format
@@ -277,13 +193,13 @@ fn parse_candump_file_msg(line: &str) -> eyre::Result<CanFrame> {
         data[i] = byte;
     }
 
-    Ok(CanFrame {
+    Ok(CanFrame::new(
         timestamp,
-        interface: interface.to_string(),
+        interface.to_string(),
         canid,
         dlc,
         data,
-    })
+    ))
 }
 
 #[cfg(test)]
@@ -294,25 +210,25 @@ mod tests {
 
     fn cli_format_fixture() -> (&'static str, CanFrame) {
         let line = "(1739136517.221471)  can0  123   [3]  0A B0 3f\n";
-        let frame = CanFrame {
-            timestamp: 1739136517.221471,
-            interface: String::from("can0"),
-            canid: 0x123,
-            dlc: 3,
-            data: [0x0A, 0xB0, 0x3F, 0, 0, 0, 0, 0],
-        };
+        let frame = CanFrame::new(
+            1739136517.221471,
+            String::from("can0"),
+            0x123,
+            3,
+            [0x0A, 0xB0, 0x3F, 0, 0, 0, 0, 0],
+        );
         (line, frame)
     }
 
     fn file_format_fixture() -> (&'static str, CanFrame) {
         let line = "(1739136482.503244) can0 123#0AB03f\n";
-        let frame = CanFrame {
-            timestamp: 1739136482.503244,
-            interface: String::from("can0"),
-            canid: 0x123,
-            dlc: 3,
-            data: [0x0A, 0xB0, 0x3F, 0, 0, 0, 0, 0],
-        };
+        let frame = CanFrame::new(
+            1739136482.503244,
+            String::from("can0"),
+            0x123,
+            3,
+            [0x0A, 0xB0, 0x3F, 0, 0, 0, 0, 0],
+        );
         (line, frame)
     }
 
@@ -355,27 +271,27 @@ mod tests {
                       (03) can0 125#0C\n
                      ";
         let expected = [
-            CanFrame {
-                timestamp: 01.0,
-                interface: String::from("can0"),
-                canid: 0x123,
-                dlc: 1,
-                data: [0x0A, 0, 0, 0, 0, 0, 0, 0],
-            },
-            CanFrame {
-                timestamp: 02.0,
-                interface: String::from("can0"),
-                canid: 0x124,
-                dlc: 1,
-                data: [0x0B, 0, 0, 0, 0, 0, 0, 0],
-            },
-            CanFrame {
-                timestamp: 03.0,
-                interface: String::from("can0"),
-                canid: 0x125,
-                dlc: 1,
-                data: [0x0C, 0, 0, 0, 0, 0, 0, 0],
-            },
+            CanFrame::new(
+                01.0,
+                String::from("can0"),
+                0x123,
+                1,
+                [0x0A, 0, 0, 0, 0, 0, 0, 0],
+            ),
+            CanFrame::new(
+                02.0,
+                String::from("can0"),
+                0x124,
+                1,
+                [0x0B, 0, 0, 0, 0, 0, 0, 0],
+            ),
+            CanFrame::new(
+                03.0,
+                String::from("can0"),
+                0x125,
+                1,
+                [0x0C, 0, 0, 0, 0, 0, 0, 0],
+            ),
         ];
         let actual: Vec<_> = CandumpParser::new(&lines[..])
             .filter_map(|m| m.ok())
@@ -390,27 +306,27 @@ mod tests {
                       (03) can0 125 [1] 0C\n\
                      ";
         let expected = [
-            CanFrame {
-                timestamp: 01.0,
-                interface: String::from("can0"),
-                canid: 0x123,
-                dlc: 1,
-                data: [0x0A, 0, 0, 0, 0, 0, 0, 0],
-            },
-            CanFrame {
-                timestamp: 02.0,
-                interface: String::from("can0"),
-                canid: 0x124,
-                dlc: 1,
-                data: [0x0B, 0, 0, 0, 0, 0, 0, 0],
-            },
-            CanFrame {
-                timestamp: 03.0,
-                interface: String::from("can0"),
-                canid: 0x125,
-                dlc: 1,
-                data: [0x0C, 0, 0, 0, 0, 0, 0, 0],
-            },
+            CanFrame::new(
+                01.0,
+                String::from("can0"),
+                0x123,
+                1,
+                [0x0A, 0, 0, 0, 0, 0, 0, 0],
+            ),
+            CanFrame::new(
+                02.0,
+                String::from("can0"),
+                0x124,
+                1,
+                [0x0B, 0, 0, 0, 0, 0, 0, 0],
+            ),
+            CanFrame::new(
+                03.0,
+                String::from("can0"),
+                0x125,
+                1,
+                [0x0C, 0, 0, 0, 0, 0, 0, 0],
+            ),
         ];
         let actual: Vec<_> = CandumpParser::new(&lines[..])
             .filter_map(|m| m.ok())
@@ -443,51 +359,5 @@ mod tests {
                         3.0,can0,0x9F8051C,2,2,0x1C,0xFF,0x1F805,0CD0\n\
                        ";
         assert_eq!(csv_str, expected);
-    }
-
-    #[test]
-    fn test_canid_parsing() {
-        let frame = CanFrame {
-            canid: 0x0CAC1C13,
-            ..Default::default()
-        };
-        assert_eq!(frame.pgn(), 0xAC00);
-        assert_eq!(frame.src(), 0x13);
-        assert_eq!(frame.dst(), 0x1C);
-
-        let frame = CanFrame {
-            canid: 0x18FF3F13,
-            ..Default::default()
-        };
-        assert_eq!(frame.pgn(), 0xFF3F);
-        assert_eq!(frame.src(), 0x13);
-        assert_eq!(frame.dst(), 0xFF);
-
-        let frame = CanFrame {
-            canid: 0x18EF1CF5,
-            ..Default::default()
-        };
-        assert_eq!(frame.priority(), 0x6);
-        assert_eq!(frame.pgn(), 0xEF00);
-        assert_eq!(frame.src(), 0xF5);
-        assert_eq!(frame.dst(), 0x1C);
-
-        let frame = CanFrame {
-            canid: 0x09F8051C,
-            ..Default::default()
-        };
-        assert_eq!(frame.priority(), 0x2);
-        assert_eq!(frame.pgn(), 0x1F805); // has the data page bit set
-        assert_eq!(frame.src(), 0x1C);
-        assert_eq!(frame.dst(), 0xFF);
-
-        let frame = CanFrame {
-            canid: 0x18EEFF1C,
-            ..Default::default()
-        };
-        assert_eq!(frame.priority(), 0x6);
-        assert_eq!(frame.pgn(), 0xEE00);
-        assert_eq!(frame.src(), 0x1C);
-        assert_eq!(frame.dst(), 0xFF);
     }
 }
